Remove interactive option in setup.sh
This makes setup.sh assume that you are running it interactively if the
remote option is not specified.

Previously the options allowed for a non-interactive local session.
However, it is currently not clear to me how that can be achieved
without compromising security, as we need to somehow store privilige
escalation credentials for the current user on the machine.