Install VS Code package from PopOS's repositories
This patch removes the usage of a role for installing VS Code.
When using a role to install VS Code, it rightfully installs the
Microsoft maintained apt repository for VS Code and subsequently the
VS Code package from it.

This is not needed for PopOS, because the package is already available
in the default repositories.

This patch also adds a different role for installing VS Code plugins.